+++ src/main/java/org/sonar/plugins/delphi/preprocessor/directive/expression/Expressions.java
@@ -16,6 +16,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
+import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.Nullable;
 import org.sonar.plugins.delphi.preprocessor.DelphiPreprocessor;
 import org.sonar.plugins.delphi.preprocessor.directive.CompilerDirective.Expression;
@@ -253,15 +254,20 @@ public class Expressions {
       return type.size();
     }
 
+    private boolean isIntrinsic(String intrinsic, int minArguments) {
+      return StringUtils.removeStartIgnoreCase(name, "System.").equalsIgnoreCase(intrinsic)
+          && arguments.size() >= minArguments;
+    }
+
     @Override
     public ExpressionValue evaluate(DelphiPreprocessor preprocessor) {
-      if (name.equalsIgnoreCase("Defined") && !arguments.isEmpty()) {
+      if (isIntrinsic("Defined", 1)) {
         Expression argument = arguments.get(0);
         if (argument instanceof NameReferenceExpression) {
           boolean isDefined = preprocessor.isDefined(((NameReferenceExpression) argument).name);
           return createBoolean(isDefined);
         }
-      } else if (name.equalsIgnoreCase("SizeOf") && !arguments.isEmpty()) {
+      } else if (isIntrinsic("SizeOf", 1)) {
         int size = sizeOf(preprocessor, arguments.get(0));
         return createInteger(size);
       }