+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for XE7 dynamic array operations:
   - See: [Dynamic Arrays in XE7](https://blog.marcocantu.com/blog/2014_september_dynamic_arrays_delphixe7.html)
+- Support for `SizeOf` intrinsic in preprocessor expressions.
 - Support for obscure `(.` and `.)` tokens:
   - These are now interpreted correctly by the lexer as square brackets.
   - See: [The Future of the Delphi Compiler](https://web.archive.org/web/20210330140547/https://edn.embarcadero.com/article/39174)